fix: cache expiring set to 1 second.
fastapi_cache only deletes when hitting expired entry so after deleting a movie it
was still appearing in get_all and get one movie info.
memcache option does not have implemented the clearing and the memcache client flush
raises conection exceptions, so the workaround is just setting an short expiring time.

Redis based option should be the optimal but requires a redis service running in the background.